+++ crates/ruma-common/src/api.rs
@@ -151,7 +151,7 @@ macro_rules! metadata {
 ///     #     rate_limited: false,
 ///     #     authentication: None,
 ///     #     history: {
-///     #         unstable => "/_matrix/some/endpoint/:room_id",
+///     #         unstable => "/_matrix/some/endpoint/{room_id}",
 ///     #     },
 ///     # };
 ///
@@ -187,7 +187,7 @@ macro_rules! metadata {
 ///     #     rate_limited: false,
 ///     #     authentication: None,
 ///     #     history: {
-///     #         unstable => "/_matrix/some/endpoint/:file_name",
+///     #         unstable => "/_matrix/some/endpoint/{file_name}",
 ///     #     },
 ///     # };
 ///