+++ crates/ruma-common/tests/api/conversions.rs
@@ -14,7 +14,7 @@ const METADATA: Metadata = metadata! {
     rate_limited: false,
     authentication: None,
     history: {
-        unstable => "/_matrix/foo/:bar/:user",
+        unstable => "/_matrix/foo/{bar}/{user}",
     }
 };
 
@@ -143,7 +143,7 @@ mod without_query {
         rate_limited: false,
         authentication: None,
         history: {
-            unstable => "/_matrix/foo/:bar/:user",
+            unstable => "/_matrix/foo/{bar}/{user}",
         }
     };